refactor: Agnostric server entry point
Created a generic entry point to avoid hard dependecy from the API framework. closes #2
